Split off RequestSource from the RequestProvider interface
