+++ src/crawlee/storages/_request_list.py
@@ -1,14 +1,12 @@
 from __future__ import annotations
 
 from collections import deque
-from datetime import timedelta
 from typing import TYPE_CHECKING
 
 from typing_extensions import override
 
 from crawlee._utils.docs import docs_group
-from crawlee.base_storage_client._models import ProcessedRequest
-from crawlee.storages._request_provider import RequestProvider
+from crawlee.storages._request_source import RequestSource
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
@@ -17,7 +15,7 @@ if TYPE_CHECKING:
 
 
 @docs_group('Classes')
-class RequestList(RequestProvider):
+class RequestList(RequestSource):
     """Represents a (potentially very large) list of URLs to crawl.
 
     Disclaimer: The `RequestList` class is in its early version and is not fully implemented. It is currently
@@ -61,10 +59,6 @@ class RequestList(RequestProvider):
         return await self.is_empty() and len(self._in_progress) == 0
 
     @override
-    async def drop(self) -> None:
-        self._requests.clear()
-
-    @override
     async def fetch_next_request(self) -> Request | None:
         try:
             request = self._requests.popleft()
@@ -91,37 +85,3 @@ class RequestList(RequestProvider):
     @override
     async def get_handled_count(self) -> int:
         return self._handled_count
-
-    @override
-    async def add_request(
-        self,
-        request: str | Request,
-        *,
-        forefront: bool = False,
-    ) -> ProcessedRequest:
-        request = self._transform_request(request)
-
-        if forefront:
-            self._requests.appendleft(request)
-        else:
-            self._requests.append(request)
-
-        return ProcessedRequest(
-            id=request.id,
-            unique_key=request.unique_key,
-            was_already_handled=False,
-            was_already_present=False,
-        )
-
-    @override
-    async def add_requests_batched(
-        self,
-        requests: Sequence[str | Request],
-        *,
-        batch_size: int = 1000,
-        wait_time_between_batches: timedelta = timedelta(seconds=1),
-        wait_for_all_requests_to_be_added: bool = False,
-        wait_for_all_requests_to_be_added_timeout: timedelta | None = None,
-    ) -> None:
-        transformed_requests = self._transform_requests(requests)
-        self._requests.extend(transformed_requests)
